renaming class and improving the code
