+++ src/main/java/com/github/syr0ws/fastinventory/api/inventory/InventoryViewer.java
@@ -6,5 +6,5 @@ public interface InventoryViewer {
 
     Player getPlayer();
 
-    InventoryHistory getInventoryHistory();
+    InventoryViewManager getViewManager();
 }
