+++ src/main/java/com/github/syr0ws/fastinventory/common/inventory/action/CloseAction.java
@@ -17,7 +17,7 @@ public class CloseAction extends CommonAction {
     @Override
     public void execute(FastInventoryClickEvent event) {
         InventoryViewer viewer = event.getViewer();
-        viewer.getInventoryHistory().close();
+        viewer.getViewManager().clear();
     }
 
     @Override
